[Feature & Refacotoring] Separate the logic for the intro page
- Error modal window when a user who is not logged in clicks the Apply button
